login page update
progress on header and login page:
Header functional, with link to sign up page.
Login page has the headings and textboxs, and links to login button and forget password.
No logic has been implemented for them yet. Their stylings are not finalized either.
Still figuring out how to get the image on the left to appear.