sign up page update
added a quick mock up for the sign up page. header for the sign up page is added as well.

The link buttons are functional, but may need to be replaced for actual buttons later.

The css sections for them have been added, but no css has been added yet.