sign up page update
completed the barebones boxes for the sign up pages.

clicking the div takes you to the sign up information form, clicking the submit will take you to the verification page, and submitting on the verification page takes you to the login page.

note: the CSSes are barebone and will be fleshed out later.

as of currently, there are no significant differences between the student sign up form and the company sign up form, besides some field names.

There are no logic implemented in the front end (no error checking
whatsoever).
There are also no names for the verification code boxes as of currently.